moving around blocks, adding in init files
