+++ examples/llm/src/layers/__init__.py
@@ -7,29 +7,14 @@ from examples.llm.src.layers.attention import (FlashCausalAttention,
                                                alibi_bias)
 from examples.llm.src.layers.flash_attention import FlashAttention, FlashMHA
 from examples.llm.src.layers.gpt_blocks import GPTMLP, GPTBlock
-from examples.llm.src.model_registry import COMPOSER_MODEL_REGISTRY
-from examples.llm.src.models.hf import (ComposerHFCausalLM, ComposerHFPrefixLM,
-                                        ComposerHFT5)
-from examples.llm.src.models.mosaic_gpt import ComposerMosaicGPT, MosaicGPT
-from examples.llm.src.tokenizer import (TOKENIZER_REGISTRY, HFTokenizer,
-                                        LLMTokenizer)
 
 __all__ = [
     'FlashAttention',
     'FlashMHA',
-    'ComposerHFCausalLM',
-    'ComposerHFPrefixLM',
-    'ComposerHFT5',
-    'COMPOSER_MODEL_REGISTRY',
     'TorchCausalAttention',
     'FlashCausalAttention',
     'TritonFlashCausalAttention',
     'alibi_bias',
     'GPTMLP',
     'GPTBlock',
-    'MosaicGPT',
-    'ComposerMosaicGPT',
-    'LLMTokenizer',
-    'HFTokenizer',
-    'TOKENIZER_REGISTRY',
 ]
