+++ examples/llm/src/models/__init__.py
@@ -1,0 +1,14 @@
+# Copyright 2022 MosaicML Examples authors
+# SPDX-License-Identifier: Apache-2.0
+
+from examples.llm.src.models.hf import (ComposerHFCausalLM, ComposerHFPrefixLM,
+                                        ComposerHFT5)
+from examples.llm.src.models.mosaic_gpt import ComposerMosaicGPT, MosaicGPT
+
+__all__ = [
+    'ComposerHFCausalLM',
+    'ComposerHFPrefixLM',
+    'ComposerHFT5',
+    'MosaicGPT',
+    'ComposerMosaicGPT',
+]
