+++ examples/llm/src/models/mosaic_gpt.py
@@ -19,54 +19,7 @@ from composer.models.base import ComposerModel
 from omegaconf import DictConfig
 
 import examples.llm.src.layers.attention as attention
-
-
-class GPTMLP(nn.Module):
-
-    def __init__(self, cfg: DictConfig, device: Optional[str] = None):
-        super().__init__()
-        self.mlp_up = nn.Linear(cfg.d_model,
-                                cfg.mlp_ratio * cfg.d_model,
-                                device=device)
-        self.mlp_act = nn.GELU(approximate='none')
-        self.mlp_down = nn.Linear(cfg.mlp_ratio * cfg.d_model,
-                                  cfg.d_model,
-                                  device=device)
-        self.mlp_down._is_residual = True  # type: ignore
-
-    def forward(self, x):
-        return self.mlp_down(self.mlp_act(self.mlp_up(x)))
-
-
-class GPTBlock(nn.Module):
-
-    def __init__(self,
-                 cfg: DictConfig,
-                 causal_attn_cls,
-                 device: Optional[str] = None):
-        super().__init__()
-        if cfg.get('alibi', False):
-            assert cfg.attn_impl == 'triton' or cfg.attn_impl == 'torch', 'Only triton kernel or torch supports alibi'
-        self.ln_1 = nn.LayerNorm(cfg.d_model, device=device)
-        self.causal_attn = causal_attn_cls(cfg, device)
-        self.ln_2 = nn.LayerNorm(cfg.d_model, device=device)
-        self.mlp = GPTMLP(cfg, device=device)
-        self.resid_attn_dropout = nn.Dropout(cfg.resid_pdrop)
-        self.resid_mlp_dropout = nn.Dropout(cfg.resid_pdrop)
-
-    def forward(
-        self,
-        x: torch.Tensor,
-        key_padding_mask: Optional[torch.ByteTensor] = None,
-        attn_mask: Optional[torch.Tensor] = None,
-    ) -> torch.Tensor:
-        a = self.ln_1(x)
-        b, _ = self.causal_attn(a, key_padding_mask, attn_mask)
-        x = x + self.resid_attn_dropout(b)
-        m = self.ln_2(x)
-        n = self.mlp(m)
-        x = x + self.resid_mlp_dropout(n)
-        return x
+import examples.llm.src.layers.gpt_blocks as gpt_blocks
 
 
 class MosaicGPT(nn.Module):
@@ -114,9 +67,9 @@ class MosaicGPT(nn.Module):
         self.transformer.update({
             'blocks':
                 nn.ModuleList([
-                    GPTBlock(cfg,
-                             causal_attn_cls=self.causal_attn_cls,
-                             device=cfg.init_device)
+                    gpt_blocks.GPTBlock(cfg,
+                                        causal_attn_cls=self.causal_attn_cls,
+                                        device=cfg.init_device)
                     for _ in range(cfg.n_layers)
                 ])
         })
@@ -322,11 +275,11 @@ class MosaicGPT(nn.Module):
 
     # FSDP Wrap function
     def fsdp_wrap_fn(self, module):
-        return isinstance(module, GPTBlock)
+        return isinstance(module, gpt_blocks.GPTBlock)
 
     # Activation Checkpointing
     def activation_checkpointing_fn(self, module):
-        return isinstance(module, GPTBlock)
+        return isinstance(module, gpt_blocks.GPTBlock)
 
 
 class ComposerMosaicGPT(ComposerModel):
